+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* Settings getting in `user-setting` and `adv-limit` services.
 
 ## [3.7.0-beta.5] - 2021-11-16
 ### Added